pyproject.toml
[project]
name = "socia-research"
version = "0.1.0"
description = "Social media research bot with connector-first architecture"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.20.0",
    "alembic>=1.13.0",
    "httpx>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0",
    "apscheduler>=3.10.0",
    "structlog>=24.1.0",
    "feedparser>=6.0.0",
    "selectolax>=0.3.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.6",
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "respx>=0.21.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]
telegram = [
    "telethon>=1.34.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
